# docs/swagger.yaml
basePath: /api/v1
definitions:
  polling-system_auth_models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  polling-system_auth_models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  polling-system_auth_models.LoginResponse:
    properties:
      expires_in:
        type: string
      id_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/polling-system_auth_models.User'
    type: object
  polling-system_auth_models.PaginatedUsers:
    properties:
      items:
        items:
          $ref: '#/definitions/polling-system_auth_models.User'
        type: array
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  polling-system_auth_models.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      firebase_uid:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  polling-system_auth_models.UserCreate:
    properties:
      display_name:
        type: string
      email:
        type: string
      firebase_uid:
        type: string
      role:
        enum:
        - admin
        - poller
        - viewer
        type: string
    required:
    - email
    - firebase_uid
    type: object
  polling-system_auth_models.UserRoleUpdate:
    properties:
      role:
        enum:
        - admin
        - poller
        - viewer
        type: string
    required:
    - role
    type: object
  polling-system_auth_models.UserUpdate:
    properties:
      display_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Unified API for authentication, polls, contacts, and file management.
  title: Polling System API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email/password via Firebase Auth, provisions
        the user if new, and returns the user profile with tokens.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_auth_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_auth_models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      summary: Login with email and password
      tags:
      - Auth
  /auth/token:
    post:
      description: Verifies the Firebase ID token (via middleware), provisions the
        user if new, and returns the user profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_auth_models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Exchange Firebase token for user profile
      tags:
      - Auth
...
  /users:
    get:
      description: Returns a paginated list of all users. Admin only.
      parameters:
      - default: 1
        description: 'Page number (min: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (1-100)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_auth_models.PaginatedUsers'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a user manually. Admin only.
      parameters:
      - description: User to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_auth_models.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/polling-system_auth_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Deletes a user. Admin only.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - Users
    get:
      description: Returns a user by database ID. Admin only.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_auth_models.User'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates a user. Admin only.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_auth_models.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_auth_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - Users
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Updates a user's role. Admin only.
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_auth_models.UserRoleUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_auth_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user role
      tags:
      - Users
  /users/me:
    get:
      description: Returns the authenticated user's profile.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_auth_models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's display name.
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_auth_models.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_auth_models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/polling-system_auth_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update own profile
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Enter your token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"